Move debug_print to jit_util
