+++ src/evox/core/__init__.py
@@ -19,9 +19,8 @@ __all__ = [
 
 # deal with vmap nesting and JIT
 from . import _vmap_fix
-from ._vmap_fix import debug_print
 from .components import Algorithm, Monitor, Problem, Workflow
-from .jit_util import jit, vmap
+from .jit_util import debug_print, jit, vmap
 
 # export symbols
 from .module import ModuleBase, Mutable, Parameter, assign_load_state_dict, jit_class, trace_impl, use_state, vmap_impl
